Big tuning improvements to pool_numa on a large scale 4-socket 192 thread system. This especially appears to improve the scaling of jemalloc.
Added a convenience script "cols.sh" for extracting pretty printed columns from the output of the microbenchmark. It is designed to make it fast and easy to extract relevant data of the form "xyz=###" when rapidly iterating over designs and performance debugging in the terminal. Similarly, added a small "grep_maxres.sh" script to extract the maxresident size from subcommands run within a "time" command.

Added some debugging info to main.cpp to better capture reasons that threads fail to spawn sometimes, e.g., when evaluating hoard with 192 threads performing 100% updates in large abtrees. (It appears to simply be caused by a flat 20000 thread creation limit in pthreads.)